2_Data_Engineering/2_Weather_Pipeline/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
matplotlib==3.8.4
